+++ Source/Core/General/Clock.cs
@@ -21,6 +21,6 @@ namespace CodeImp.DoomBuilder
 	public static class Clock
 	{
 		// This queries the system for the current time
-		public static float CurrentTime { get { return Configuration.Timer.ElapsedMilliseconds; } }
+		public static long CurrentTime { get { return Configuration.Timer.ElapsedMilliseconds; } }
 	}
 }
